refactor(sim_codegen): extract gen_linklist into submodule
